Corporation buf for PB88.
This commit changes the output amount of corporations and
reduce the headquarter income from 4 to 3.

Note this commit also contain my currently disabled proposal:
The yield of corporations sinks if the locations spreads in the own
nation. This can be debufed by spreading in foreign countries.

To enable this, change CORPORATION_PBMOD_MINIMAL_FACTOR in
GlobalDefinesAlt.xml and rename CvCorporationScreen.py.disabled