Update ScoreBoard and Ctrl+R at game loading
Add Interface function to call some commands 'without widget'
This also to simulates 'Ctrl+R' at game start.

Moreover this commit fixes an logical error in CvEventManager.py:
The code in onUpdate()... will not be called because this handler
is not activated in this Mod.
I shift the code into onGameUpdate.