PBSpy: migrate subcribed_user field
This ManyToManyField is now an own class, MailSubscribing(models.Model),
and was expanded by the field 'watched_player_id'.
If the player with above id finished his turn, the user will be
informed by email.
This is useful for sequential PitBoss games and wars with spit turns.

0017_auto_20200622_2128.py will copy the content of the old M2M field
into the new one. The following migrations remove the old field and
rename the new one...