PB Mod: Add better chat functionality for Pitboss host.
The normal chat function PB.sendChat(cp1252_str) had a few drawbacks.
It added a long prefix and the chat messages will be stored many rounds
in the log. This is to verbose.

This commit adds a method to send strings over multiple onModNetMessage
to clients. The clients combine the message parts and showing them
without any logging.
Moreover, some sound events can be triggered to catch the players
attention.