Enable advanced start for szenario files.
This changes allow the advanced start for szenarios. Units
and cities which are placed in the WBSave will not be removed.
Moreover, you could no logout and re-login without loosing the
advanced start screen.

The automatic conversion from start points into gold was disabled.
You need now to hit the 'Start' button twice to leave the advanced
start mode.